Update signal.py
I have been reading up the deepdive documentation. Are there more examples of composition code in the actual source code?
I watched the video on fronted API
https://www.youtube.com/watch?v=kTj5mMvN5DM&list=PLwNuX3xB_tv-wTpVDMSJr7XW6IP_qZH0t&index=4
An example is given on how to implement the concatenate method but in this case i presume the hann_window is of a different case.