Refactor Frontend tests. (#9692)
Update `handle_frontend_test` decorator to pass `test_flags` object, to minimize the number of exposed parameters and to be able to scale with no changes required to be done to all tests in the frontends. This was previously done to `handle_test` in commit eade4aa, and refactored all frontend tests to adhere to the new signature.

Revert `test_frontend_function` to test for only one function, moved the `all_aliases` to the decorator to be sampled during test execution.