Added an if statement to check whether there are any changes to stash before doing so.
This version checks whether there are any changes to stash before doing so, and uses git stash save instead of git stash to provide a description of the stash. It also checks whether there is a stash to pop before doing so, to avoid errors if there are no stashes to apply.