updated minkowski funntion
as you told I added shortened function  root_w is a variable used to store the root of the weight w. I used a dictionary to store the values of root_w for different values of p. If p is not 1, 2, or infinity.  use the ** operator to calculate the root of w.and I use list comprehension to calculate the weighted difference vector, applying the square root of w to each element in u_v if w is not None.