Update devcontainer.json
name: Specifies the name of the development environment. In this case, it is "Ivy Development Environment (image)".

image: Defines the Docker image to be used for the development environment. Here, it uses the "unifyai/ivy:latest" image.

customizations: Contains customizations specific to the Visual Studio Code (VSCode) editor. In this example, it installs the "ms-python.python" extension.

postCreateCommand: Specifies the commands to be executed after the container is created. The "post_create" command runs the "post_create_commands.sh" script, and the "bashrc" command adds an alias for the Python command in the user's bashrc file.

initializeCommand: Defines the command to pull the Docker image before creating the container. Here, it pulls the "unifyai/ivy:latest" image.

forwardPorts: Allows specifying a list of ports inside the container that should be forwarded to the local machine. Currently commented out.

runArgs: Uncommented when using a ptrace-based debugger like C++, Go, or Rust. It adds necessary permissions for debugging within the container.

mounts: Uncommented to use the Docker