Update discrete_fourier_transform.py
Fixed alignment issue and included statement to check if s shape and array shape are the same. It is difficult to pad and crop on the frontend as it results in inconsistent results between ivy.ifftn and numpy.fft.ifftn, RECOMMENDED : Code an ivy.ifftn backend function using numpy to allow for cropping and padding when the array shape and s argument shape do not match.