In the process of working on the fmod function, it was observed that this function, similar to its numpy counterpart, does not accept complex numbers as inputs. This was causing the test cases to fail. To address this, a function named get_numeric_dtype_without_complex was created to exclude complex numbers when running this specific function.
From a code reusability perspective, it might not be the best practice to place this function next to the fmod testing function, as it could potentially be useful elsewhere. However, in terms of code readability, having it close to the fmod testing function is beneficial, as it provides immediate context about its purpose and usage.