Add handle_complex_input to documentation
Add the new decorator to the function_wrapping deep dive document, as
well as modify the list to be automatically numbered (which will
hopefully keep `git diff` sizes smaller in future). Also add a link to
the doc-builder repo in the building_the_docs_pipeline deep dive, so
future contributors don't suffer as I have.