Refactor paddle backend pow and exp
To improve support of complex numbers. pow was having stability issues
due to exp(log(r)), while exp had no native complex number support so
this was added (and eventually the call to it from pow was dropped, but
its still useful to have). Also adjusted the tests of both, and of gelu