fix: resolve device mismatch in where with scalars
Previously, the `ivy.where` function would fail on the torch backend
when it was called with one (CPU) array input and one scalar, since the
scalar would be automatically placed on the GPU by
`handle_array_like_without_promotion`. This commit modifies that wrapper
so that, if a function is passed at least one array, all new arrays
created by the wrapper will be placed on the same device as that array.